% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_inv.R
\name{add_inv}
\alias{add_inv}
\title{This function returns the inverse of x under addition mod n.}
\usage{
add_inv(x, n)
}
\arguments{
\item{x}{an integer}

\item{n}{an integer}
}
\value{
x-inverse under addition mod n
}
\description{
This function returns the inverse of x under addition mod n.
}
\examples{
add_inv(2, 10)
}

