% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mod_n.R
\name{add_mod_n}
\alias{add_mod_n}
\title{This function performs addition mod n}
\usage{
add_mod_n(a, b, n)
}
\arguments{
\item{a}{an integer}

\item{b}{an integer}

\item{n}{an integer}
}
\value{
a + b mod n
}
\description{
This function performs addition mod n
}
\examples{
add_mod_n(2,16,10)
}

