% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateLimiter.R
\name{append_timelog}
\alias{append_timelog}
\title{Append timelog}
\usage{
append_timelog()
}
\value{
timestamp
}
\description{
Appends new entry to the timelog environment
 on execution of \code{log_rate_limit()}.
}
\examples{
\dontrun{
append_timelog()
}
}
