% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary2integer.R
\name{binary2integer}
\alias{binary2integer}
\title{Get integer from binary representation}
\usage{
binary2integer(x)
}
\arguments{
\item{x}{A vector that corresponds to the binary representation
of an integer. The left most element is the most significant bit.}
}
\description{
This function gets the corresponding integer from the given binary
representation.
}
