% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_reverse.R
\name{bit_reverse}
\alias{bit_reverse}
\title{Bit-reverse copy of polynomial coefficients for FFT}
\usage{
bit_reverse(a)
}
\arguments{
\item{a}{A vector of polynomial coefficients from the smallest to the
highest power. That is, a=(a_0,a_1, ..., a_n-1).}
}
\description{
This function returns the bit-reverse copy of polynomial coefficients
for the use of efficient FFT implementation. For example, given a
polynomial's coefficients a=(1,2,3,4), this function will return
A=(1,3,2,4).
}
