% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateLimiter.R
\name{check_timelog}
\alias{check_timelog}
\title{Check Timelog}
\usage{
check_timelog()
}
\value{
integer
}
\description{
Checks records added to the log environment in the
last 60 seconds for use in rate limit checks
}
\examples{
\dontrun{
check_timelog()
}
}
