% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cmc_api}
\alias{cmc_api}
\title{CoinMarketCap Professional API Call}
\usage{
cmc_api(url)
}
\arguments{
\item{url}{CoinMarketCap API URL}
}
\value{
api results
}
\description{
CoinMarketCap Professional API Call
}
