% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_code_fun.R
\name{create.code}
\alias{create.code}
\title{creates a secret code}
\usage{
create.code(key)
}
\arguments{
\item{key}{An integer number}
}
\description{
This function creates the code for decyphering the encryption system, consisting in a dataset with alpha-numerical values and the corresponding encryption. Different keys produce different codes.
}
\examples{
code1 <- create.code(key = 123)
code1
}
\keyword{code}
\keyword{decryption}
\keyword{encryption}
\keyword{key}

