% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/decrypt_fun.R
\name{decrypt}
\alias{decrypt}
\title{decrypts a message}
\usage{
decrypt(message, key)
}
\arguments{
\item{message}{An encrypted message}

\item{key}{An integer number}
}
\description{
This function decrypts a text message based on a certain reference number, called "key". The "key" needs to be shared by the sender with the recipient for him/her to be able to decrypt the message.
}
\examples{
decrypted.msg <- decrypt(message = ".92.37.109.26.97.37.109.53", key = 123)
decrypted.msg
}
\keyword{decryption}
\keyword{encryption}
\keyword{key}

