% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt_affine.R
\name{decrypt_affine}
\alias{decrypt_affine}
\title{Decrypts an affine cryptosystem.}
\usage{
decrypt_affine(data)
}
\arguments{
\item{data}{list object with a, b, and cyphertext entries}
}
\value{
character clear text string
}
\description{
Decrypts an affine cryptosystem.
}
\examples{
decode_affine(data)
}

