% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt_polyalphabetic.R
\name{decrypt_polyalphabetic}
\alias{decrypt_polyalphabetic}
\title{Decrypts an polyalphabetic cryptosystem.}
\usage{
decrypt_polyalphabetic(data)
}
\arguments{
\item{data}{list object with A, b, and cyphertext entries}
}
\value{
character cleartext string
}
\description{
Decrypts an polyalphabetic cryptosystem.
}
\examples{
decrypt_polyalphabetic(data=list(A=matrix(c(3,1,5,2), ncol=2), b=c(2,2),cyphertext="RRGR"))
}

