% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffie_hellman_key_exchange.R
\name{diffie_hellman_key_exchange}
\alias{diffie_hellman_key_exchange}
\title{Diffie-Hellman key exchange}
\usage{
diffie_hellman_key_exchange(p, g, a, b)
}
\arguments{
\item{p}{A large prime number that is used to define the group of units \eqn{\mathbb{F}_p^*} in the finite filed \eqn{F_p}. In practice, p should be
a prime of at least 1000 bits to maintain the security of the key exchange procedure.}

\item{g}{An integer in \eqn{\mathbb{F}_p^*}. For various of reasons, an ideal g should have a large prime order in the filed. Some guidelines suggest
g's order is approximately p/2.}

\item{a}{A secret integer used by Alice to calculate \eqn{g^a} modulo p.}

\item{b}{A secret integer used by Bob to calculate \eqn{g^b} modulo p.}
}
\description{
This function implements the Diffie-Hellman key exchange procedure. The security of this procedure is based on the hardness assumption of the
discrete logarithm problem. This is not a complete cryptosystem because it only allows two parties to exchange a secrete key on a public
channel securely.
}
