% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptic_curve_addition.R
\name{elliptic_curve_addition}
\alias{elliptic_curve_addition}
\title{Elliptic curve addition}
\usage{
elliptic_curve_addition(A, x1, y1, x2 = NULL, y2 = NULL)
}
\arguments{
\item{A}{The coefficient of X in the elliptic curve E.}

\item{x1}{The x-coordinate of the point P.}

\item{y1}{The y-coordinate of the point P.}

\item{x2}{The x-coordinate of the point Q. The default is NULL, indicating P=Q.}

\item{y2}{The y-coordinate of the point Q. The default is NULL, indicating P=Q.}
}
\description{
This function adds two points P and Q on an elliptic curve E: Y^2 = X^3 + AX + B. The two points are not symmetric against the x-axis.
None of them is the "origin" of E. If the coordinates of Q are not given, then the function calculates the sum P+P=2P.
}
