% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/encrypt_fun.R
\name{encrypt}
\alias{encrypt}
\title{encrypts a message}
\usage{
encrypt(message, key)
}
\arguments{
\item{message}{A text message}

\item{key}{An integer number}
}
\description{
This function encrypts a text message based on a certain reference number, called "key". The "key" needs to be shared with the recipient for him/her to be able to decrypt the message.
}
\examples{
encrypted.msg <- encrypt(message = "Ush ush.", key=123)
encrypted.msg
}
\keyword{crypto,}
\keyword{encrypt}

