% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt_affine.R
\name{encrypt_affine}
\alias{encrypt_affine}
\title{Encrypts an affine cryptosystem.}
\usage{
encrypt_affine(data)
}
\arguments{
\item{data}{list object with a, b, and cyphertext entries}
}
\value{
character cyphertext string
}
\description{
Encrypts an affine cryptosystem.
}
\examples{
encrypt_affine(data=list(a=5,b=3,cleartext="ALGEBRA"))
}

