% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt_polyalphabetic.R
\name{encrypt_polyalphabetic}
\alias{encrypt_polyalphabetic}
\title{Encrypts a polyalphabetic cryptosystem.}
\usage{
encrypt_polyalphabetic(data)
}
\arguments{
\item{data}{list object with A, b, and cleartext entries}
}
\value{
character cyphertext string
}
\description{
Encrypts a polyalphabetic cryptosystem.
}
\examples{
encrypt_polyalphabetic(data=list(A=matrix(c(3,1,5,2), ncol=2), b=c(2,2),cleartext="HELP"))
}

