% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{fft}
\alias{fft}
\title{Fast Fourier Transform}
\usage{
fft(a)
}
\arguments{
\item{a}{A vector of polynomial coefficients from the smallest to the
highest power. That is, a=(a_0,a_1, ..., a_n-1).}
}
\description{
Fast Fourier Transform (FFT) that gets sample representation of
polynomials from their coefficient representations. This is an
efficient iterative FFT implementation as described on page 917 of
the book "Introduction to Algorithms" 3rd edition by Thomas Cormen
et al.
}
