% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft_polynomial_convolution.R
\name{fft_polynomial_convolution}
\alias{fft_polynomial_convolution}
\title{Polynomial convolution}
\usage{
fft_polynomial_convolution(a, b, N)
}
\arguments{
\item{a}{A vector of polynomial coefficients. The coefficients
must be stored in a in the order of (a0,a1, ..., an).}

\item{b}{A vector of polynomial coefficients. The coefficients
must be stored in a in the order of (b0,b1, ..., bn).}

\item{N}{An integer that ensures the returned polynomial has
degree at most N - 1.}
}
\description{
Given two polynomials (stored in their coefficients), this function
returns the polynomial convolution product. More precisely, it first
calculates the product of the given two polynomials, then take
modulo x^N - 1, where N is an integer. The returned polynomial has
degree at most N - 1.
}
