% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft_polynomial_multiplication.R
\name{fft_polynomial_multiplication}
\alias{fft_polynomial_multiplication}
\title{Efficient polynomial multiplication using FFT}
\usage{
fft_polynomial_multiplication(a, b)
}
\arguments{
\item{a}{A vector of polynomial coefficients.}

\item{b}{A vector of polynomial coefficients.}
}
\description{
This function does efficient polynomial multiplication using FFT.
The two inputs are coefficients of two polynomials of arbitrary
degrees. The function relies on FFT to transform coefficient
representations to sample representations, then does multiplication
pointwise and transform back to coefficient representation of the
final answer.
}
