% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.R
\name{gcd}
\alias{gcd}
\title{Greatest common divisor}
\usage{
gcd(a, b)
}
\arguments{
\item{a}{An integer.}

\item{b}{An integer.}
}
\description{
This function uses the Euclidean algorithm to find the gcd of two integers.
}
