% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_helpers.R
\name{get_coinlist_api}
\alias{get_coinlist_api}
\title{Get Updated Coinlist}
\usage{
get_coinlist_api()
}
\value{
dataframe
}
\description{
Gets updated coin listing from CoinMarketCap professional API
and prompts user to provide CMC API key using the keyring package.
}
\examples{
\dontrun{
get_coinlist_api()
 }
}
