% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_helpers.R
\name{get_coinlist_static}
\alias{get_coinlist_static}
\title{Get Updated Coinlist}
\usage{
get_coinlist_static()
}
\value{
dataframe
}
\description{
Gets updated coin listing from local data
}
\examples{
\dontrun{
get_coinlist_static()
 }
}
