% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoins.R
\name{getCoins}
\alias{getCoins}
\title{getCoins Get historical market listings}
\usage{
getCoins()
}
\value{
Returns a data frame with over 650k rows for your pleasure.
 \tabular{lccclcccl}{
   symbol \tab  \tab  \tab  \tab chr \tab  \tab  \tab  \tab KIN \cr
   date \tab  \tab  \tab  \tab date \tab  \tab  \tab  \tab 17/12/17 \cr
   open \tab  \tab  \tab  \tab dbl \tab  \tab  \tab  \tab 0.000088 \cr
   high \tab  \tab  \tab  \tab dbl \tab  \tab  \tab  \tab 0.000137 \cr
   low \tab  \tab  \tab  \tab dbl \tab  \tab  \tab  \tab 0.000083 \cr
   close \tab  \tab  \tab  \tab dbl \tab  \tab  \tab  \tab 0.000125 \cr
   volume \tab  \tab  \tab  \tab dbl \tab  \tab  \tab  \tab 557972 \cr
   market \tab  \tab  \tab  \tab dbl \tab  \tab  \tab  \tab 66221100 \cr
   name \tab  \tab  \tab  \tab chr \tab  \tab  \tab  \tab Kin \cr
   ranknow \tab  \tab  \tab  \tab int \tab  \tab  \tab  \tab 104 \cr
}
}
\description{
This is the main function of this package and once run
will go and scrape all the historical tables of all the
different cryptocurrencies listed on CoinMarketCap and
turn it into a dataset.
}
\examples{
\dontrun{
markets <- getCoins()
}
}
