% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrices.R
\name{getPrices}
\alias{getPrices}
\title{Get current crypto currency prices}
\usage{
getPrices(coin = NULL, limit = 0, currency = NULL)
}
\arguments{
\item{coin}{Token name, default is all, Default: NULL}

\item{limit}{Return top n coins, default is all, Default: 0}

\item{currency}{Convert into local currency. Must be one of "AUD", "BRL", "CAD", "CHF", "CLP", "CNY", "CZK", "DKK", "EUR", "GBP", "HKD", "HUF", "IDR", "ILS", "INR", "JPY", "KRW", "MXN", "MYR", "NOK", "NZD", "PHP", "PKR", "PLN", "RUB", "SEK", "SGD", "THB", "TRY", "TWD", "ZAR", Default: NULL}
}
\value{
Will provide data frame of current prices
}
\description{
This will retrieve the current market prices from CoinMarketCap. Data gets refreshed every 5 minutes.
}
\details{
Updated every 5 minutes
}
\examples{
{
kin_price <- getPrices("kin")
}
}
