% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexadecimal2decimal.R
\name{hexadecimal2decimal}
\alias{hexadecimal2decimal}
\title{Hexadecimal to decimal}
\usage{
hexadecimal2decimal(x)
}
\arguments{
\item{x}{An integer.}
}
\description{
This function convers hexadecimal to decimal integer representation.
}
