% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateLimiter.R
\name{initiate_timelog}
\alias{initiate_timelog}
\title{Initiate Timelog Environment}
\usage{
initiate_timelog()
}
\value{
environment
}
\description{
Initiates environment for storage of time logs for rate limit checks
}
\examples{
\dontrun{
initate_timelog()
}
}
