% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_helpers.R
\name{install_keyring}
\alias{install_keyring}
\title{CoinMarketCap Professional API Call}
\usage{
install_keyring()
}
\description{
Checks to see if keyring package is installed
and then if not prompts for installation
}
