% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer2binary.R
\name{integer2binary}
\alias{integer2binary}
\title{Binary integer representation}
\usage{
integer2binary(x)
}
\arguments{
\item{x}{An integer.}
}
\description{
This function returns the binary representation of an integer. The output is a vector, in which the left most element is the most significant digit
and the right most element is the least significant digit.
}
