% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer_modular_multiplicative_inverse.R
\name{integer_modular_multiplicative_inverse}
\alias{integer_modular_multiplicative_inverse}
\title{Modular inverse}
\usage{
integer_modular_multiplicative_inverse(a, p)
}
\arguments{
\item{a}{An integer that is less than p and coprime with p.}

\item{p}{An integer that is coprime with a.}
}
\description{
This function uses the exteneded Euclidean algorithm to find the
inverse of an integer a modulo p. The inverse only exists if a and
p are coprime. If F_p is a field, then all none zero elements in it
has multiplicative inverse.
}
