% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_divisible.R
\name{is_divisible}
\alias{is_divisible}
\title{Divisibility}
\usage{
is_divisible(a, b)
}
\arguments{
\item{a}{An integer.}

\item{b}{A non-zero integer that is smaller than a.}
}
\description{
This function checks if an integer is divisible by another integer.
}
