% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listCoins.R
\name{listCoins}
\alias{listCoins}
\title{Retrieve Cryptocurrencies}
\usage{
listCoins()
}
\value{
Returns data frame of coin listings for getting
 coin names, rank and slugs for use in \code{getCoins()}
 and populating the urls for use in \code{scraper()}.
 \tabular{lccclcccl}{
  symbol \tab  \tab  \tab  \tab chr \tab  \tab  \tab  \tab "KIN"\cr
  name \tab  \tab  \tab  \tab chr \tab  \tab  \tab  \tab "Kin"\cr
  slug \tab  \tab  \tab  \tab chr \tab  \tab  \tab  \tab "kin"\cr
  rank \tab  \tab  \tab  \tab dbl \tab  \tab  \tab  \tab 104\cr
}
}
\description{
This is called automatically as part of \code{getCoins()}.
It retrieves name, slug, symbol and rank of cryptocurrencies from
CoinMarketCap and creates URLS for \code{scraper()} to use.
}
\examples{
\dontrun{
coin_list <- listCoins()
}
}
