% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_mul_inv.R
\name{matrix_mul_inv}
\alias{matrix_mul_inv}
\title{This function returns the modular inverse of 2-by-2 matrix ().}
\usage{
matrix_mul_inv(A, n)
}
\arguments{
\item{A}{a 2-by-2 matrix}

\item{n}{an integer}
}
\value{
modular inverse of matrix A
}
\description{
This function returns the modular inverse of 2-by-2 matrix ().
}
\examples{
matrix_mul_inv(A, 26)
}

