% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miller_rabin_primarily_test.R
\name{miller_rabin_primarily_test}
\alias{miller_rabin_primarily_test}
\title{Primality test}
\usage{
miller_rabin_primarily_test(n, a)
}
\arguments{
\item{n}{An integer to be tested for primality.}

\item{a}{Another integer, called a Miller-Rabin witness for the compositeness of n.}
}
\description{
This function uses the Miller-Rabin test for composite numbers. The test will either return TRUE indicates
n is a composite number of NULL indicates unsure.
}
