% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular_exponentiation.R
\name{modular_exponentiation}
\alias{modular_exponentiation}
\title{Modular exponentiation}
\usage{
modular_exponentiation(b, e, m)
}
\arguments{
\item{b}{An integer base.}

\item{e}{An integer exponent.}

\item{m}{An integer modulus.}
}
\description{
This function calculates modular exponentiation in an (space) efficient way.
}
