% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular_exponentiation_binary.R
\name{modular_exponentiation_binary}
\alias{modular_exponentiation_binary}
\title{Modular exponentiation}
\usage{
modular_exponentiation_binary(b, e, m)
}
\arguments{
\item{b}{An integer base.}

\item{e}{An integer exponent.}

\item{m}{An integer modulus.}
}
\description{
This function calculates modular exponentiation using the square-and-multiply method. It is faster than the function modular_exponentiation().
}
