% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular_square_root.R
\name{modular_square_root}
\alias{modular_square_root}
\title{Modular square root}
\usage{
modular_square_root(a, p)
}
\arguments{
\item{a}{An integer whose square root exists modulo the prime p.}

\item{p}{A prime modular.}
}
\description{
This function calculates the square root of an integer modulo a prime that is congruent to 3 modulo 4. This method is valid only if the given
integer has a square root modulo the chosen prime.
}
