% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monic_polynomial.R
\name{monic_polynomial}
\alias{monic_polynomial}
\title{Monic polynomial}
\usage{
monic_polynomial(a, p)
}
\arguments{
\item{a}{A vector of polynomial coefficients.}

\item{p}{A integer. The polynomial coeffiicents are taken from Z/pZ.
If p is not prime, the function may fail, because the leading term
coefficient may not be coprime with p.}
}
\description{
This function makes a polynomial to monic. That is, the coefficient
of the largest degree term is 1. This can help to simplify
polynomial devision.
}
