% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mul_mod_n.R
\name{mul_mod_n}
\alias{mul_mod_n}
\title{This function performs multiplication mod n}
\usage{
mul_mod_n(a, b, n)
}
\arguments{
\item{a}{an integer}

\item{b}{an integer}

\item{n}{an integer}
}
\value{
a * b mod n
}
\description{
This function performs multiplication mod n
}
\examples{
mul_mod_n(3,16,10)
}

