% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omit_zero.R
\name{omit_zero}
\alias{omit_zero}
\title{Omit zeros}
\usage{
omit_zero(a)
}
\arguments{
\item{a}{A vector of polynomial coefficients.}
}
\description{
Given a polynomial that is stored in its coefficient format, there may
be 0s at the back of the coefficient vector. This function omits
zeros at the back of the vector, so the length of the vector is the
polynomial degree + 1.
}
