% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_helpers.R
\name{platform_locale}
\alias{platform_locale}
\title{Platform Locale}
\usage{
platform_locale()
}
\value{
locale
}
\description{
Helper function to identify locale to use for date encoding
based on source platform being unix based or windows
}
\examples{
{
platform_locale()
}
}
