% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_addition.R
\name{polynomial_addition}
\alias{polynomial_addition}
\title{Polynomial addition}
\usage{
polynomial_addition(a, b)
}
\arguments{
\item{a}{A vector of polynomial coefficients. The coefficients
must be stored in a in the order of (a0,a1, ..., an).}

\item{b}{A vector of polynomial coefficients. The coefficients
must be stored in a in the order of (b0,b1, ..., bn).}
}
\description{
Given two polynomials (stored in their coefficient vectors), this
function returns the sum of them. Summation is component wise. The
two polynomials can be of same or different degrees.
}
