% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_degree.R
\name{polynomial_degree}
\alias{polynomial_degree}
\title{Polynomial degree}
\usage{
polynomial_degree(a)
}
\arguments{
\item{a}{A vector of polynomial coefficients, from the smallest term
to the largest term.}
}
\description{
This function returns the degree of a polynomial.
}
