% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_gcd.R
\name{polynomial_gcd}
\alias{polynomial_gcd}
\title{Greatest common divisor between polynomials}
\usage{
polynomial_gcd(a, b, p)
}
\arguments{
\item{a}{A vector of polynomial coefficients.}

\item{b}{A vector of polynomial coefficients.}

\item{p}{A integer. The polynomial coeffiicents are taken from Z/pZ.
If p is not prime, the algorithm may fail, because some integer
coefficients may not have inverses in Z/pZ, causing polynomial
division to fail.}
}
\description{
This function uses the Euclidean algorithm to find the gcd of two polynomials.
Since we are always interested in integer coefficient polynomials,
we require coefficients to be taken from Z/pZ. Ideally, p is prime.
}
