% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_modular_multiplicative_inverse.R
\name{polynomial_modular_multiplicative_inverse}
\alias{polynomial_modular_multiplicative_inverse}
\title{Modular inverse}
\usage{
polynomial_modular_multiplicative_inverse(a, N, p)
}
\arguments{
\item{a}{A polynomial represented by its coefficients. It should be
coprime with the polynomial x^N-1. It's degree should be smaller than
N. If not, a pre-processing step is neede before running this function.}

\item{N}{An integer for the degree of the modulus polynomial X^N-1.}

\item{p}{An prime integer such that polynomial coefficients are
taken from Z/pZ.}
}
\description{
Given a polynomail from a polynomial quotient ring R_p=Z_p[x]/(x^N - 1),
this function uses the exteneded Euclidean algorithm to find the
inverse of the polynomial modulo in R_p. For p prime, the inverse
exists if and only if the given polynomial and X^N-1 are coprime.
}
