% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_modulo.R
\name{polynomial_modulo}
\alias{polynomial_modulo}
\title{Polynomial modulo}
\usage{
polynomial_modulo(a, N)
}
\arguments{
\item{a}{A vector of polynomial coefficients. The coefficients
must be stored in a in the order of (a0,a1, ..., an).}

\item{N}{An integer that ensures the returned polynomial has
degree at most N - 1.}
}
\description{
Given a polynomial (stored in its coefficients), this function
returns the polynomial modulo x^N - 1, where N is an integer.
}
