% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clue.R
\name{read_clue}
\alias{read_clue}
\title{Reads a .clue formated file.}
\usage{
read_clue(filepath)
}
\arguments{
\item{filepath}{string containing relative path to .clue file}
}
\value{
list containing data from the file
}
\description{
Reads a .clue formated file.
}
\examples{
read_clue("../clues/filename.clue")
}

