% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{repair_dependencies}
\alias{repair_dependencies}
\title{Installs all dependant packages}
\usage{
repair_dependencies()
}
\value{
dependant packages
}
\description{
Helper function to install dependencies
}
\examples{
\dontrun{
# Fix dependency issues by reinstalling all
fix_crypto <- repair_dependencies()
}
}
