% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_helpers.R
\name{replace_encoding}
\alias{replace_encoding}
\title{Check locale encoding}
\usage{
replace_encoding(sys_locale)
}
\arguments{
\item{sys_locale}{string system locale}
}
\value{
Sets system variable to UTF-8
}
\description{
Helper function to ensure encoding is UTF-8
}
\examples{
\dontrun{
sys_locale <- Sys.getlocale(category = "LC_TIME")
replace_encoding(sys_locale)
}
}
