% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_helpers.R
\name{safely_read_json}
\alias{safely_read_json}
\title{Safely read json API}
\usage{
safely_read_json(json_url)
}
\arguments{
\item{json_url}{json API URL}
}
\value{
parsed dataset or a elegant error message
}
\description{
This will attempt to safely retrieve data and return a elegant error message to the user if the data is unavailable.
}
\examples{
{
json_url <- 'https://s2.coinmarketcap.com/generated/search/quick_search.json'
result <- safely_read_json(json_url)
}

}
