% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{scraper}
\alias{scraper}
\title{Scraping Historical Tables from CMC}
\usage{
scraper(attributes, slug)
}
\arguments{
\item{attributes}{Singular url input from \code{listCoins()}}

\item{slug}{coin from \code{listCoins()}#}
}
\value{
Returns data.frame of scraped results.
 \tabular{lccclcccl}{
  Date \tab  \tab  \tab  \tab chr \tab  \tab  \tab  \tab "Dec 17, 2017"\cr
  Open \tab  \tab  \tab  \tab dbl \tab  \tab  \tab  \tab 0.000088\cr
  High \tab  \tab  \tab  \tab dbl \tab  \tab  \tab  \tab 0.000137\cr
  Low \tab  \tab  \tab  \tab dbl \tab  \tab  \tab  \tab 0.000083\cr
  Close \tab  \tab  \tab  \tab dbl \tab  \tab  \tab  \tab 0.000125\cr
  Volume \tab  \tab  \tab  \tab chr \tab  \tab  \tab  \tab 557,972\cr
  Market \tab  \tab  \tab  \tab chr \tab  \tab  \tab  \tab 66,221,100\cr
  symbol \tab  \tab  \tab  \tab chr \tab  \tab  \tab  \tab "KIN"\cr
}
}
\description{
Retrieves historical tables from CoinMarketCap for the input token.
This function is a dependency of \code{getCoins()}
and is used as part of the loop to retrieve all cryptocurrencies.
}
\examples{
\dontrun{
scraper(attributes)
}
}
