% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateLimiter.R
\name{use_rate_limit}
\alias{use_rate_limit}
\title{Use Rate Limit}
\usage{
use_rate_limit(call, limit = 30, verbose = FALSE)
}
\arguments{
\item{call}{Function to receive and apply rate limit on.}

\item{limit}{Integer for rate limit per minute. DEfault is 30.}

\item{verbose}{Logical to display progress messaging. Default is FALSE}
}
\value{
Expected output of function called
}
\description{
Apply rate limiter to function call which
stores variables inside an environment on execution to
ensure rate limits on API requests are adhered to.
}
\examples{
\dontrun{
fun_function <- function(x,y) (x+y) / y
use_rate_limit(fun_function(6+16), 10, verbose = FALSE)
}
}
