% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_add_inv.R
\name{vector_add_inv}
\alias{vector_add_inv}
\title{This function returns the inverse of vector x  under addition mod n.}
\usage{
vector_add_inv(x, n)
}
\arguments{
\item{x}{a vector of length 2}

\item{n}{an integer}
}
\value{
x-inverse under addition mod n
}
\description{
This function returns the inverse of vector x  under addition mod n.
}
\examples{
vector_add_inv(c(2,2), 10)
}

