% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_polynomial.R
\name{zero_polynomial}
\alias{zero_polynomial}
\title{Check zero polynomial}
\usage{
zero_polynomial(a)
}
\arguments{
\item{a}{A vector of polynomial coefficients.}
}
\description{
This function checks if a given polynomial is a zero polynomial.
}
