% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_helpers.R
\name{cmc_api}
\alias{cmc_api}
\title{CoinMarketCap Professional API Call}
\usage{
cmc_api(url)
}
\arguments{
\item{url}{CoinMarketCap API URL}
}
\value{
api results
}
\description{
CoinMarketCap Professional API Call
}
\examples{
\dontrun{
cmc_api("https://pro-api.coinmarketcap.com/v1/cryptocurrency/listings/latest")
}
}
