% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_history.R
\name{crypto_history}
\alias{crypto_history}
\title{Get historic crypto currency market data}
\usage{
crypto_history(coins = NULL, limit = NULL, start_date = NULL,
  end_date = NULL, coin_list = NULL, sleep = NULL)
}
\arguments{
\item{coins}{string if NULL retrieve all currently existing coins (crypto_list()),
or provide list of crypto currencies in the crypto_list() format (e.g. current and dead coins since 2015)}

\item{limit}{integer Return the top n records, default is all tokens}

\item{start_date}{string Start date to retrieve data from, format 'yyyymmdd'}

\item{end_date}{string End date to retrieve data from, format 'yyyymmdd'}

\item{coin_list}{string Valid values are 'api', 'static' or NULL}

\item{sleep}{integer Seconds to sleep for between API requests}
}
\value{
Crypto currency historic OHLC market data in a dataframe and additional information via attribute "info":
  \item{slug}{Coin url slug}
  \item{symbol}{Coin symbol}
  \item{name}{Coin name}
  \item{date}{Market date}
  \item{open}{Market open}
  \item{high}{Market high}
  \item{low}{Market low}
  \item{close}{Market close}
  \item{volume}{Volume 24 hours}
  \item{market}{USD Market cap}
  \item{close_ratio}{Close rate, min-maxed with the high and low values that day}
  \item{spread}{Volatility premium, high minus low for that day}
  \item{start_date}{in info: Begin of historic data}
  \item{end_date}{in info: End of historic data}
  \item{message}{in info: Either "Success" when data was available or error message from scraper}

This is the main function of the crypto package. If you want to retrieve
ALL coins then do not pass a argument to crypto_history(), or pass the coin name.
}
\description{
Scrape the crypto currency historic market tables from
CoinMarketCap <https://coinmarketcap.com> and display
the results in a date frame. This can be used to conduct
analysis on the crypto financial markets or to attempt
to predict future market movements or trends.
}
