% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{fft}
\alias{fft}
\title{Fast Fourier Transform}
\usage{
fft(a, inverse = FALSE)
}
\arguments{
\item{a}{A vector of polynomial coefficients from the smallest to the
highest power (or samples when inverser is set to TRUE).
That is, a=(a_0,a_1, ..., a_n-1). Ideally, the length
of a is a power of 2. If not, it will be packed with 0s at the back
to make it so. Hence, the output will be a vector of length n that
is a power of 2.}

\item{inverse}{A logical parameter. The default is FALSE. When set
to TRUE, the function will calculate the inverse of FFT.}
}
\description{
Fast Fourier Transform (FFT) that gets sample representation of
polynomials from their coefficient representations. This is an
efficient iterative FFT implementation as described on page 917 of
the book "Introduction to Algorithms" 3rd edition by Thomas Cormen
et al.
}
